junior csharp migralas
